aarch64: Relax aarch64_simd_mem_operand_p
In the testcase, the arm_neon.h expansion code was generating
intrinsics whose mems had (subreg:DI (reg:CADI …) 0) addresses,
but these addresses were then rejected by the underlying patterns.

The problem was that the define_expands required the address
to be a pmode_register_operand (and thus allowed subregs) whereas
aarch64_simd_mem_operand_p required a REG or a POST_INC address
(and so disallowed subregs).  IMO it was the latter behaviour
that was wrong, since the rules for what is a valid base register
should be the same here as elsewhere.