Fix various other &MEM_REF folds
A previous patch fixed a loose &MEM_REF -> POINTER_PLUS_EXPR
fold in expand_simple_operations.  This patch tries to fix
possible instances of the same problem elsewhere.

These changes are purely by inspection, so (a) I don't have any
testcases that prove that the changes are needed and (b) there
are likely to be other instances of the same problem that I missed.
But I think the changes make sense on first principles, so I'd
prefer to err on the side of having them rather than not.

I tried various ways of adding a helper function for this, but all
uses are just about different enough that the abstraction felt more
awkward than useful.