c: Propagate capability-ness through &addr->foo etc.
This patch makes expressions like:

  &addr->foo
  &addr[i]
  &*addr
  etc.

have capability type whenever addr has capability type.  It also
makes array-to-pointer conversion return a capability type whenever
the base address of the array is a capability, such as in:

  cap->array + 0

The patch adds a c_build_addr_expr function that enforces these
rules and (defensively) uses the function for all instances of:

  build_unary_op (..., unqualified_addr_expr (), ...)

even those (like omp) that we're not in a position to test properly.

I wondered about putting some of this in c-family, but I think the
C++ version is likely to look significantly different.  The code is
also quite tightly tied to the C-specific lvalue_p function.

Clang agrees on the assertions in the testcase, except for the
CAP ones for __real and __imag lvalues.  I think that's a bug
in clang though.