Factor sync-builtins.def definitions
Adding _c support to the sync builtins in their current form
would have meant an awful lot of cut-&-paste.  I wasn't confident
I could do that without introducing inconsistencies, so this patch
first refactors sync-builtins.def so that each group of functions
has one top-level definition.

This doesn't come for free.  The indirection means that it's
harder at a glance to tell what the signature of each group of
functions actually is.  On the other hand, it does become easier
to see at a glance what the groups are.

I think on balance the patch should help to make the code more
maintainable.