Tweak prototypes of __atomic_fetch_*_capability
At the function prototype level, the capability forms of the
following functions treated “val” as a capability:

- __atomic_fetch_<operation>(ptr, val, memorder)
- __atomic_<operation>_fetch(ptr, val, memorder)
- __sync_fetch_and_<operation>(ptr, val)
- __sync_<operation>_and_fetch(ptr, val)

However, these arguments are used to add to, subtract from,
or otherwise do arithmetic on the capability data.  The type
of the argument for __atomic_fetch_add should therefore be
the type associated with “*ptr + val”, etc.

The same argument applies to non-capability pointer types,
and I think it's a bug that GCC (for all targets) accepts:

  int *foo(int **ptr, int *val) {
    return __atomic_fetch_add(ptr, val, __ATOMIC_ACQUIRE);
  }

without warning.  Clang in contrast says:

  warning: incompatible pointer to integer conversion passing 'int *' to parameter of type 'long' [-Wint-conversion]

But I think this becomes more important with capabilities,
since if we view the functions as operations on two capabilities,
we introduce ambiguity about provenance.  It also affects the
type signature and ABI of any out-of-line implementations.

This patch therefore makes the “val” arguments above have
non-capability type, matching clang's behaviour.  The code
generated by the functions already (correctly) dropped the
capability, so this is purely a change to the prototype rather
than a change to the underlying behaviour.

At the same time, the patch optimises the implementation of the add
functions so that they can use ADD Cn, Cn, Xn rather than a separate
ADD + SCVALUE.  (In particular, the split happens later than combine,
so we can't rely on that pass to do the optimisation.)