aarch64: Optimise atomic stores of TI
Atomic TI stores must be 16-byte aligned.  This means that,
on Morello, we can implement them using:

	st...	czr, [base]

This should be better even for normal-base addresses.  However,
for alternative-base addresses it has the additional benefit
of avoiding an out-of-line function, since there are no
alternative-base forms of the atomic pair instructions that
TI operations normally use.