aarch64: Make aarch64_atomic_exchange<mode> require !TARGET_LSE
aarch64_atomic_exchange<mode>_lse comes after
aarch64_atomic_exchange<mode> and contains the same pattern,
just with fewer clobbers.  This doesn't matter when we first
expand the instructions, since we generate a pattern without
clobbers for LSE and a pattern with clobbers otherwise.

However, if combine finds a valid optimisation later, it will
try the patterns in order and pick the first one that matches.
This matching process ignores trailing clobbers, since combine
can add those itself where necessary.

The net result is that combine can end up turning an LSE
operation into a slower non-LSE one.  I don't have a standalone
testcase for this, but it was shown up by tests in later patches.

Three possible fixes are:

(1) Differentiate the main part of the patterns (e.g. by using
    different unspecs).

(2) Swap the patterns around.

(3) Make aarch64_atomic_exchange<mode> require !TARGET_LSE.

(1) feels like a hack, and would be a non-starter if we had direct rtl
codes for representing the operation.  (2) would be OK, but the ordering
requirement is subtle enough that it would need a comment.  That would
leave us having to explain away why we allow both patterns for LSE,
but choose an order in which the second will never actually be used
(for LSE) in practice.

The patch therefore goes for (3).