lra: Fix reloading of addresses in atomic ops [PR105395]
If the result of an atomic operation becomes a hard register
before reload, and if that hard register conflicts with the
allocated address of the atomic operation, we can end up reloading
the whole atomic memory instead of the address.  I saw this
“in the wild” with some later changes, because a combine opportunity
on the input triggered a three-way combination with a following
hard-register move.

The problem was that lra-constraints.c only set offmemok if the memory
didn't “win” (i.e. if the memory didn't satisfy the constraints in its
original form).  The memory here is OK in its original form.  Later on
the function detected the conflict and marked the memory as no longer
winning:

          if (HARD_REGISTER_P (operand_reg[i])
              || (first_conflict_j == last_conflict_j
                  && operand_reg[last_conflict_j] != NULL_RTX
                  && !curr_alt_match_win[last_conflict_j]
                  && !HARD_REGISTER_P (operand_reg[last_conflict_j])))
            {
              curr_alt_win[last_conflict_j] = false;

But because offmemok was false, this led to a reload of the full
memory rather than its address.