Add alternative-base support for __atomic_load, etc.
This patch adds alternative-base support for the following
overloaded atomic functions:

  __atomic_load
  __atomic_store
  __atomic_exchange
  __atomic_compare_exchange

sync-builtins.def gave these functions a specific prototype,
but the prototype had the wrong number of arguments: it started
with a size argument that the functions don't actually have.
The patch therefore gives the functions a BT_FN_<RET>_VAR type
instead, for consistency with the overloaded _n functions.

Otherwise, the only substantial change is to make sure that the first
pointer argument to the resolved call has the same “flavour” as the
first pointer argument to the original call.  The second argument to
__atomic_compare_exchange is always a default-flavour pointer, even
for the _c version.  (This is also what clang does.)