Fix creation of TARGET_MEM_REFs with __capability alias pointers
TARGET_MEM_REFs use the offset operand to convey two pieces
of information: the alias type for TBAA purposes and a constant
offset from the base.  create_mem_ref_raw was using fold_convert
to merge the alias type with the integer offset, but that doesn't
work when the alias type is a __capability pointer.