Update the dl_iterate_phdr implementation
A hosted unwinder uses dl_iterate_phdr in order to find the location of
exception handling information.  The structure that this callback
receives for Morello now contains a capability.  A local internal
structure needs to be updated to accomodate this.

The existing code works for most cases.  This is because the pointer to
the exception handling frame section which is found during this
iteration has been given metadata to span the entire executable by the
kernel and libc.

There is one exceptional case.  When the load_base of the executable is
zero (i.e. for a PDE or statically linked executable) it is not always
possible for the kernel and libc to provide a capability with zero as
the value and tight bounds spanning the relevant binary.  For this case
we derive a pointer to the frame using a found address and metadata from
the dl_phdr_info->dlpi_phdr member, which is set up to span the
executable.

In order to improve robustness we also add some extra assertions that
metadata is valid and spans relevant addresses.  The placement of these
assertions are taken from Morello-LLVM's libunwind in order to catch the
same sort of problems that Morello LLVM has been adjusted to look for.