aarch64: Avoid generating PC-relative stores
c675754a1788ddd741fb310d8d7ee612b95264f5 regressed -mcmodel=tiny
by allowing PC-relative stores such as “str x0, foo”, whereas
only PC-relative loads are supported.  Previously
aarch64_classify_address had:

      if (ldp_stp_mode == VOIDmode
	  && GET_MODE_SIZE (mode).is_constant (&const_size)
	  && const_size >= 4)
	{
	  rtx sym, addend;

	  split_const (x, &sym, &addend);
	  return (GET_CODE (sym) == LABEL_REF
		   || (GET_CODE (sym) == SYMBOL_REF
		       && aarch64_pcrelative_literal_loads
		       && aarch64_function_literal_pool_address_p (x)));
	}

and so allowed references to labels (which should never be needed here)
and references to function literal pools.  The patch instead made the
code test whether x is SYMBOL_TINY_ABSOLUTE, which is needed for
-mcmodel=tiny references to .capinit entries in .data.rel.ro.

However, as the fallout shows, we still need the test for a constant
pool entry.  It just doesn't need to be in the function literal pool.

As the comment in the patch says, this is really a hack.  The
pre-Morello handling of -mcmodel=tiny means that we generate:

	adr	x0, foo
	ldr	x0, [x0]

even though:

	ldr	x0, foo

would be OK too.  But fixing things so that we generate that
without:

	str	x0, foo

would be too much of a distraction.  It's also something that
should happen on trunk first.