morello: Tidy up defines and enable Morello-specific ones in Hybrid
This patch resolves a couple of the MORELLO TODOs we had in the codebase to
remove the Morello defines from fake-capability compilation.

It also removes `__CHERI_CAP_PERMISSION_PERMIT_CCALL__` to match the ACLE
and ensures that the remaining ones are also defined when -march=morello.
This now matches clang behaviour.