testsuite: Add cheri_capability_any effective target
As of Stam's recent patch, we only define __CHERI__ and friends for real
capability targets (not fakecap). This allows us to define an effective
target that only matches for hybrid and purecap (not fakecap).

Apply the filter to a couple of tests that should no longer be run on
fakecap with the recent macro changes.