aarch64: Add hybridcap multilib
This adds a multilib for hybrid-capability Morello. This enables execution
testing of hybrid Morello with a GCC-built baremetal toolchain.

When building newlib for hybrid with GCC, we ran into the issue of
cheri.h using __builtin_cheri_cap_{to,from}_pointer, but these being
as-yet unimplemented in GCC. To get around this for now, we comment out
the definitions of the wrapper functions for these intrinsics in
cheri.h.