testsuite: Extend aapcs64 tests to hybrid
This patch adjusts the aapcs64 tests to also run on hybrid. We don't add
any new tests with this patch, this simply adjusts existing tests to
also work on hybrid.

We adjust the abitest* code to save the capability registers on hybrid
(as we do for purecap) and add some __capability annotations to make
existing purecap tests also valid for hybrid.

There remain issues with the tests involving packed structures (these
fail even on purecap) that this patch doesn't address.

Note that we remove target selectors from the morello-* tests since (as
aapcs64.exp already notes) these are ignored when using c-torture-execute, and
aapcs64.exp already ensures these tests are run on the correct targets.

It is left for follow-on patches to add any hybrid-specific tests (e.g.
mixing capability and integer pointers, passing capability varargs
parameters, etc.)