avoid early reference to debug-only symbol
If some IPA pass replaces the only reference to a constant non-public
non-automatic variable with its initializer, namely the address of
another such variable, the former becomes unreferenced and it's
discarded by symbol_table::remove_unreachable_nodes.  It calls
debug_hooks->late_global_decl while at that, and this expands the
initializer, which assigs RTL to the latter variable and forces it to
be retained by remove_unreferenced_decls, and eventually be output
despite not being referenced.  Without debug information, it's not
output.

This has caused a bootstrap-debug compare failure in
libdecnumber/decContext.o, while developing a transformation that ends
up enabling the above substitution in constprop.

This patch makes reference_to_unused slightly more conservative about
such variables at the end of IPA passes, falling back onto
expand_debug_expr for expressions referencing symbols that might or
might not be output, avoiding the loss of debug information when the
symbol is output, while avoiding a symbol output only because of debug
info.


for  gcc/ChangeLog

	* dwarf2out.c (add_const_value_attribute): Return false if
	resolve_one_addr fails.
	(reference_to_unused): Don't assume local symbol presence
	while it can still be optimized out.
	(rtl_for_decl_init): Fallback to expand_debug_expr.
	* cfgexpand.c (expand_debug_expr): Export.
	* expr.h (expand_debug_expr): Declare.