Handle some special cases for bitregion_end in bitwise operations
We recently tested a patch for mode-wise mentioning of strict alignment
by artificially regression testing a patch which claimed that SImode
values had to be loaded from alignment boundaries but all other modes
did not.
In this regression test we hit two ICE's from the compiler going into an
infinite loop between extract_fixed_bit_field and
extract_split_bit_field.  This was tracked down to a clause in
expand_expr_real_1 setting bitregion_end and bitregion_start to values
that caused trouble later on.

These values represent the start bit position of an access inside a
structure or array (bitregion_start) and an upper bound of the end of
that access (bitregion_end).  In expansions of tree nodes other than
COMPONENT_REF accesses the bitregion_end is calculated using the
MEM_SIZE of the containing data type (e.g. entire structure or entire
array).  This is only done when MEM_SIZE is known, otherwise we fall
back to a default of an unknown region size.

The two edge cases we came across were a zero-sized array in pr72802.c,
and a zero-sized array member of a structure in pr45819.c.

    #pragma pack(1)
    struct S0 {
      short f3;
      float f4;
      signed f5
    };
    const struct S0 t7[] = {};

    struct ehci_regs {
	char x;
	unsigned int port_status[0];
    } __attribute__ ((packed));

In these cases our calculation of bitregion_end became less than zero,
and less than bitregion_start respectively.  In the case bitregion_end
was less than bitregion_start bitregion_end became less than zero in an
adjustment inside extract_bit_field.

When extract_fixed_bit_field was given a negative bitregion_end, this
was passed down to get_best_mode and from there to
bit_field_mode_iterator::next_mode.  This function checks if we can
access a bitfield with a given mode by checking if the size of the
access goes past the end of the bitregion.  A negative bitregion_end
appeared to be crossed by any access.

The fix is to fall back to an unknown bitregion in such cases.  This
seems to fit with the goal of the bitwise checks in the "zero-sized
array member" access, since such arrays should still be valid to access
(and we would have to rely on the capability hardware to catch
out-of-bounds accesses here).  Also, as mentioned below, accesses to an
integer mode as a bitfield are just done in that integer mode for
Morello, and we can't have an accessible zero-width bitfield in order to
try and trigger the "zero sized structure member access" case.

N.b. this was only triggered by this artificial compiler state.
Specifying that SImode required alignment meant that we attempted to
access the above members `port_status` and `f5` through bitfields (as
they were unaligned SImode values).  For capability values this does not
trigger because the bitwise access is done in integer modes (which means
that extract_bit_field_1 loads the value with a TImode access rather than
go through extract_fixed_bit_field to hit this bug.

Testcases that should trigger this problem on a relevant target are in
the testsuite, and there is no way of triggering this problem for the
Morello target.  Hence we do not add any tests.