libstdc++: two hacks to get the purecap build going
This commit disables two static_asserts in libstdc++ that fail when
building libstdc++ for purecap. This is a compromise to enable progress
for Morello toolchain development.

In one case (in cow-stdexcept.cc) we turn the build-time failure into a
runtime failure by substituting the static_assert for __builtin_abort.
In memory_resource.cc, we simply remove the static_assert. This is a
known area of libstdc++ that will need porting work for Morello.

This enables us to build a full C++ baremetal cross toolchain for
Morello (with a purecap multilib).