Change advertised version of the Morello/CHERI compiler.
As it stands the Morello/CHERI GCC compiler is based off of a point
mid-way in development towards GCC 11.

This is not ideal, and we intend to rebase onto a release branch when we
have time.

Another more immediate problem is that the compiler reports it is GCC-11
when this is not correct.  We do not support most of the features that
GCC-11 added from GCC-10.

This can be confusing, and has already caused problems with version
checks in glibc header files.

This patch adjusts our advertised version to be GCC-10 to improve the
imperfect situation.