Hybrid Morello: Add warnings on pointer conversions and an error on global capability initialisation
This patch adds warning messages on implicit and
explicit conversions between integer pointers and
capability pointers (both ways).

These warnings are NOT fully consistent with clang,
as we do not yet implement __chari_fromcap and
__cheri_tocap on the casts to/from capabilities.
When those are implemented we can update the wording
of the warnings and the -W<option> flags to be in
line with clang.
For now, I simply opted to be as verbose as possible
on the warnings and -W<option> flags.

Additionally, I've added an error message for the
initialisation of a global capability with a non
capability constant value. This kind of casting
is not supported and it will in the future be enabled
only via the capability-address-of operand.