c: Fix double evaluation of intcap expressions
For an intcap c and address offset cv, the C FE would generate:

REPLACE_ADDRESS_VALUE (c, (long) c + cv)

but this is problematic if c cannot safely be evaluated twice (e.g.
for volatile variables, side-effecting function calls, etc.) as we end
up getting any side effects twice. This can be seen with the
intptr-eval.c test added with this patch, which was miscompiled on
purecap Morello.

To fix this, we need to use SAVE_EXPR on the intcap from which
provenance is carried when building the REPLACE_ADDRESS_VALUE in the FE,
so we get the tree:

REPLACE_ADDRESS_VALUE (SAVE_EXPR <c>, (long) SAVE_EXPR <c> + cv)

instead. This also means we need to make the provenance decision up
front, so we know which operand needs SAVE_EXPRs. We add a new helper
function, binary_op_get_intcap_provenance, to do this. We add this in
c-family/c-common.c in anticipation that the same code might be useful
in the C++ FE. Note that this change in how we make provenance decisions
renders c_build_replace_address_value_1 and its wrapper functions
redundant, so we remove these.

As well as binary operations, similar fixes are needed for
build_unary_op and its subroutine intcap_increment. In build_unary_op,
for TRUTH_NOT_EXPR, we allow the intcap to be handled directly by the
normal code (i.e. without dropping the capability): we don't want to
re-derive an intcap here, and intcaps can be handled just fine by
c_objc_common_truthvalue_conversion.

Adding these SAVE_EXPRs can prevent some folding of
REPLACE_ADDRESS_VALUE that we were getting before, thereby regressing
code quality. The SAVE_EXPRs have TREE_SIDE_EFFECTs set, so some folds
that were possible before are no longer possible. One such example is:

(long) RAV (c, cv) --> cv.

This transformation is only valid if c is free of side effects. In
particular, if c is a SAVE_EXPR, we can't do this transformation.
However, when this expression is gimplified, any SAVE_EXPRs will be
lowered. At this point we should be able to apply the transformation.
Hence, we add a match.pd rule for this that can be applied to the gimple
after the SAVE_EXPRs are lowered. This fixes a code quality regression
seen with nested binary ops such as:

__intcap f(__intcap x, long y) { return x + y + y; }

where (without the match.pd rule) we would end up generating two scvalue
instructions. Further issues with the folding have been identified as
part of this work, but are left to a follow-on patch to address.

One further adjustment we make is to save_expr, where we teach it to
push the save_expr inside REPLACE_ADDRESS_VALUE calls (down to the
operands).  This allows the SAVE_EXPRs to be eliminated entirely in some
cases (e.g.  when the operands are integer constants) and in general
improves our chances of being able to fold the resulting expressions.

gcc/c-family/ChangeLog:

	* c-common.c (binary_op_get_intcap_provenance): New.
	(unary_op_get_intcap_provenance): New.
	* c-common.h (binary_op_get_intcap_provenance): New.
	(unary_op_get_intcap_provenance): New.

gcc/c/ChangeLog:

	* c-typeck.c (c_build_replace_address_value_1): Delete.
	(c_build_replace_address_value): Delete.
	(c_fold_build_replace_address_value): Delete.
	(intcap_increment): Avoid evaluating any intcap side-effects
	twice.
	(build_unary_op): Likewise.
	(build_binary_op_1): Likewise.

gcc/ChangeLog:

	* match.pd (non-cap) RAV (c, cv) -> (non-cap) cv): New.
	* tree.c (save_expr): Push save_expr inside
	REPLACE_ADDRESS_VALUE operands.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/morello/intcap-semantics.c: Add SAVE_EXPRs
	to some cases, following FE adjustments.
	* gcc.c-torture/execute/intptr-eval.c: New test.