middle-end: Fix bogus folding of REPLACE_ADDRESS_VALUE
fold_build_replace_address_value_loc did the following transformation:

RAV (RAV (c, cv1), cv2) -> cv2

which is incorrect in the case where cv1 has side effects. It also
(incorrectly) attempted to handle a similar case where the inner RAV was
inside a NOP_EXPR.

In the general case, both of these folds really seem to be optimizations
that should be handled by the gimple optimizers rather than
transformations that are necessary for constant folding in the FE.

The folds we do in fold_build_replace_address_value_loc should really be
the minimum necessary for constant folding in the FE. Any further
optimization can be done on GIMPLE using match.pd rules.

We adjust the nested RAV fold in fold_build_replace_address_value_loc to
only handle the case where the inner RAV and the new capability value
are both TREE_CONSTANT.

This should be sufficient for FE constant folding (e.g. for the new test
added with the patch), but without implementing any further
optimization. This should also ensure we fix the issue where cv1 has
side effects, since in this case the inner RAV would not be marked
TREE_CONSTANT.

This revealed that we weren't marking IFN CALL_EXPR nodes as
TREE_CONSTANT even if both operands were TREE_CONSTANT, so we fix this
by adjusting tree.c:process_call_operands.  Here we mark IFN calls as
TREE_CONSTANT if they are flagged ICF_CONST and ICF_NOTHROW, and all of
their operands are also TREE_CONSTANT. This further revealed that we
hadn't given REPLACE_ADDRESS_VALUE IFNs the ECF_NOTHROW flag, so we add
this flag in internal-fn.def.

We add a match.pd rule to pick up the general optimization of nested
REPLACE_ADDRESS_VALUE calls at the GIMPLE level. This prevents
regressing now that we have restricted the generic folding to do the
bare minimum for constant folding.

We remove an overzealous assert from
varasm.c:initializer_constant_valid_p_1. Here we attempted to enforce
the invariant that we never see nested REPLACE_ADDRESS_VALUE calls, but
it simply isn't possible to maintain this invariant in the general case,
since these can't always be folded, so we simply reject these instead.

gcc/ChangeLog:

	* fold-const.c (fold_build_replace_address_value_loc): Only do
	bare minimum needed for constant folding, thereby avoiding a
	correctness problem where we lose side effects.
	* internal-fn.def (REPLACE_ADDRESS_VALUE): Add ECF_NOTHROW.
	* match.pd (RAV (RAV (c, cv1), cv2) -> RAV(c, cv2)): New.
	* tree.c (process_call_operands): Propagate TREE_CONSTANT for
	IFNs marked ECF_CONST and ECF_NOTHROW.
	* varasm.c (initializer_constant_valid_p_1): Don't assert on
	nested REPLACE_ADDRESS_VALUE.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/morello/intcap-global-const.c: New test.