except.c: Fix up switching to .data.rel.ro for purecap
C++ templates use COMDAT groups to keep all codegen artifacts related to
a single template instantiation together, so they can all be discarded
or kept together.

On purecap Morello, we introduced a switch to the .data.rel.ro section
which happens halfway through outputting the .gcc_except_table section.
However, this switch to .data.rel.ro didn't preserve any COMDAT group
for the current_function_decl, so the .chericap we would insert in the
.data.rel.ro section wasn't cleaned up by the linker's COMDAT group
handling. This could lead to "dangerous relocation: out of range" errors
as we would have stale R_MORELLO_CAPINIT relocations being applied to
incompatible different instantiations of the same template.

Notably the switch_to_exception_section function already has logic to
apply the COMDAT group for the current function to the .gcc_except_table
section we switch to, so this patch reuses this logic to do the same
when we switch to the .data.rel.ro section.

gcc/ChangeLog:

	* except.c (switch_to_relro_section): New. Use it ...
	(dw2_output_call_site_table): ... here.
	(get_eh_named_section): New. Use it ...
	(switch_to_exception_section): ... here.