Handle overloading of __builtin_cheri functions
This patch adds overloading of __builtin_cheri functions, using the
approach described above resolve_cheri_builtin.  Initially I'd tried
some other approaches, the first of which meant having to move the
built-in definitions from aarch64 to generic code.  That probably
wasn't necessary with the final version, but given that I'd done
all the work before switching to that, it seemed worth keeping
anyway.

(The earlier versions tried to make the whole IL polymorphic,
to reduce the amount of casting, but that got quite ugly.)

Testing this for C++ on hybrid showed a couple of places where
we weren't preserving the “capability-ness” of pointers.

I'll follow up with a patch to move the remaining non-overloaded
functions to target-independent code too (just for consistency,
not because it's inherently better).