Make bit_field_mode_iterator behave conservatively for capabilities
bit_field_mode_iterator assumes that increasing an access from
M bytes to N bytes won't trap whenever (a) the access is known
to be aligned to N bytes and (b) N is not in danger of approaching
a likely page size.  This assumption isn't safe for capabilities,
since without any information to the contrary, we have to assume
that increasing the size of an access could take it out of bounds.

This patch adds an extra parameter to the bit_field_mode_iterator
constructor that indicates whether the address is a capability.
The alignment optimisation then requires this new parameter
to be false.

Providing nonzero bitregion_start and bitregion_end values
for extractions thus becomes an optimisation, rather than
something that is needed for correctness.  Zero values work
conservatively correctly while nonzero values give
bit_field_mode_iterator the information it needs to expand
the access in a safe way.

The bitregion parameters therefore become redundant if their
only purpose was to prevent the access being widened on capability
targets (as for emit_group_load_1).  Also, the bitregions used
for extracts in store_split_bit_field and store_integral_bit_field
weren't right, since the values applied to the target rtx (op0)
rather than the source rtx (value).

The get_inner_reference change is needed to avoid a regression
in pr69990.c.  Without it we would access the bitfield in that
testcase according to -fstrict-volatile-bitfields rules, even
though the structure is smaller than the underlying type of
the bitfield (see PR105805).

Using the alignment of the innermost type is supposed to be
a conservative fix, in terms of changing the volatile bitfield
semantics as little as possible.  It means that for things like:

struct S {
  __attribute__((packed)) volatile int a : 16;
  volatile _Float16 b; // byte offset 2
  int c; // gives the structure 4-byte alignment
};

we'll continue to use a 32-bit access for "a" (thus loading "b" too).
The handling continues to be asymmetrical: accessing "b" will not
access "a" in the same way.

The store_split_bit_field and extract_split_bit_field changes
are because aligned word-sized accesses are no longer guaranteed
to be allowed.  Without the changes there would be infinite
recursion between the split and fixed functions.